+++ packages/jira-adapter/test/filters/script_runner/workflow/workflow_link_type.test.ts
@@ -16,10 +16,10 @@
 import { filterUtils } from '@salto-io/adapter-components'
 import { InstanceElement, toChange } from '@salto-io/adapter-api'
 import _ from 'lodash'
-import { createEmptyType, getFilterParams, mockClient } from '../../utils'
-import orFilter from '../../../src/filters/script_runner/workflow/workflow_lists_parsing'
-import { WORKFLOW_TYPE_NAME } from '../../../src/constants'
-import { getDefaultConfig } from '../../../src/config/config'
+import { createEmptyType, getFilterParams, mockClient } from '../../../utils'
+import orFilter from '../../../../src/filters/script_runner/workflow/workflow_lists_parsing'
+import { WORKFLOW_TYPE_NAME } from '../../../../src/constants'
+import { getDefaultConfig } from '../../../../src/config/config'
 
 describe('ScriptRunner linkTypes in DC', () => {
   let filter: filterUtils.FilterWith<'onFetch' | 'preDeploy' | 'onDeploy'>
